feat: generate a json file to reflect tier list for future games
