+++ scripts/extractors/index.ts
@@ -13,3 +13,4 @@ export { extractAndSavePastGamesToFeeds } from "./feeds";
 export { extractAndSaveTierListGames } from "./tier-list-games";
 export { extractAndSaveTierListBacklog } from "./tier-list-backlog";
 export { extractAndSaveTierListCategories } from "./tier-list-categories";
+export { extractAndSaveTierListGamesFuture } from "./tier-list-games-future";
